Add ContentRoutingMany interface.
We had been using the ProvideMany method in some Routing implementations to
improve performance.

We had several places where we needed to create this same interface to
cast Routers to check if they support the ProvideMany method. To avoid
that, and to have a common source of truth we think we should make this
interface live with the other Routing ones.